python/examples/fastmcp/requirements.txt
fastmcp
python-dotenv
lightspark
aiohttp
